extracted editcoopinfo from coopinfo component
