+++ src/components/AddHouseForm.vue
@@ -276,7 +276,6 @@ export default defineComponent({
       this.$emit('cancel-addHouseModal');
     },
     resetHouseDataFields(houseData: any) {
-      console.log('yes');
       for (let field in houseData) {
         if (typeof houseData[field] === 'object') {
           this.resetHouseDataFields(houseData[field]);
